requirements.txt
PyYAML
numpy
pyStixel-lib
torch
einops
